+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/messages/MessageComposerViewModel.kt
@@ -33,6 +33,11 @@ public class MessageComposerViewModel(
     public val input: MutableStateFlow<String> = messageComposerController.input
 
     /**
+     * If the message will be shown in the channel after it is sent.
+     */
+    public val alsoSendToChannel: MutableStateFlow<Boolean> = messageComposerController.alsoSendToChannel
+
+    /**
      * Represents the remaining time until the user is allowed to send the next message.
      */
     public val cooldownTimer: MutableStateFlow<Int> = messageComposerController.cooldownTimer
@@ -58,6 +63,12 @@ public class MessageComposerViewModel(
     public val commandSuggestions: MutableStateFlow<List<Command>> = messageComposerController.commandSuggestions
 
     /**
+     * Current message mode, either [MessageMode.Normal] or [MessageMode.MessageThread]. Used to determine if we're sending a thread
+     * reply or a regular message.
+     */
+    public val messageMode: MutableStateFlow<MessageMode> = messageComposerController.messageMode
+
+    /**
      * Gets the active [Edit] or [Reply] action, whichever is last, to show on the UI.
      */
     public val lastActiveAction: Flow<MessageAction?> = messageComposerController.lastActiveAction
@@ -68,6 +79,14 @@ public class MessageComposerViewModel(
      * @param value Current state value.
      */
     public fun setMessageInput(value: String): Unit = messageComposerController.setMessageInput(value)
+
+    /**
+     * Called when the "Also send as a direct message" checkbox is checked or unchecked.
+     *
+     * @param alsoSendToChannel If the message will be shown in the channel after it is sent.
+     */
+    public fun setAlsoSendToChannel(alsoSendToChannel: Boolean): Unit =
+        messageComposerController.setAlsoSendToChannel(alsoSendToChannel)
 
     /**
      * Called when the message mode changes and the internal state needs to be updated.